+++ example/lib/pages/welcome_page.dart
@@ -2,6 +2,7 @@ import 'dart:io';
 
 import 'package:better_player_example/pages/basic_player_page.dart';
 import 'package:better_player_example/pages/cache_page.dart';
+import 'package:better_player_example/pages/controller_controls_page.dart';
 import 'package:better_player_example/pages/controls_configuration_page.dart';
 import 'package:better_player_example/pages/event_listener_page.dart';
 import 'package:better_player_example/pages/hls_subtitles_page.dart';
@@ -98,6 +99,9 @@ class _WelcomePageState extends State<WelcomePage> {
       }),
       _buildExampleElementWidget("Memory player", () {
         _navigateToPage(MemoryPlayerPage());
+      }),
+      _buildExampleElementWidget("Controller controls page", () {
+        _navigateToPage(ControllerControlsPage());
       }),
     ];
   }